# swagger.yaml
openapi: 3.0.0
info:
  title: Scientific Publications API
  version: 1.0.0
  description: |
    A comprehensive backend API for analyzing scientific publications with NLP and ML capabilities.
    This system provides endpoints for managing publications, authors, citations, topics, and analytics.
    
    ## Features
    - Publication management with full-text search
    - Author profiles with collaboration networks
    - Citation analysis and relationship mapping
    - Topic modeling and trend analysis
    - Advanced analytics and insights
    - Data import/export capabilities
    - Secure authentication and authorization
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Limits vary by endpoint type.
  contact:
    name: API Support
    email: support@scientific-publications.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /import/upload:
    post:
      tags:
        - Import/Export
      summary: Upload document file
      description: Uploads and processes a document file (PDF, DOC, DOCX, TXT)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                authors:
                  type: string
                year:
                  type: integer
                keywords:
                  type: string
                abstract:
                  type: string
      responses:
        '200':
          description: File uploaded and processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  file:
                    $ref: '#/components/schemas/FileInfo'
                  result:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
  /import/csv:
    post:
      tags:
        - Import/Export
      summary: Import from CSV
      description: Imports publications from a CSV file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                batchSize:
                  type: integer
                  default: 50
                updateExisting:
                  type: boolean
                  default: false
      responses:
        '200':
          description: CSV import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  file:
                    type: string
                  result:
                    $ref: '#/components/schemas/ImportResult'
  /import/json:
    post:
      tags:
        - Import/Export
      summary: Import from JSON
      description: Imports publications from a JSON file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                batchSize:
                  type: integer
                  default: 50
                updateExisting:
                  type: boolean
                  default: false
      responses:
        '200':
          description: JSON import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  file:
                    type: string
                  result:
                    $ref: '#/components/schemas/ImportResult'
  /import/batch:
    post:
      tags:
        - Import/Export
      summary: Batch import multiple files
      description: Imports multiple files in batch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
                batchSize:
                  type: integer
                  default: 50
                updateExisting:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Batch import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  files:
                    type: array
                    items:
                      type: object
                  result:
                    type: object
  /import/extract-info:
    post:
      tags:
        - Import/Export
      summary: Extract structured information from document
      description: Extracts structured data from uploaded document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Information extracted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    $ref: '#/components/schemas/FileInfo'
                  extracted_info:
                    type: object
  /import/validate:
    post:
      tags:
        - Import/Export
      summary: Validate import data
      description: Validates data structure before import
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
                - format
              properties:
                data:
                  type: object
                format:
                  type: string
                  enum: [json, csv]
      responses:
        '200':
          description: Data validation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_records:
                    type: integer
                  valid_records:
                    type: integer
                  invalid_records:
                    type: integer
                  validation_errors:
                    type: array
                    items:
                      type: object
                  is_valid:
                    type: boolean
  /import/status:
    get:
      tags:
        - Import/Export
      summary: Get import status and configuration
      description: Retrieves current import service status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Import status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  import_service:
                    type: object
                  etl_service:
                    type: object
  /import/etl/trigger:
    post:
      tags:
        - Import/Export
      summary: Trigger manual ETL
      description: Manually triggers ETL processes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sources:
                  type: array
                  items:
                    type: string
                    enum: [openalex, crossref]
                  default: [openalex, crossref]
                fullSync:
                  type: boolean
                  default: false
      responses:
        '200':
          description: ETL process started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sources:
                    type: array
                    items:
                      type: string
                  full_sync:
                    type: boolean
                  result:
                    type: object
  /import/etl/status:
    get:
      tags:
        - Import/Export
      summary: Get ETL job status
      description: Retrieves status of ETL jobs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ETL status retrieved
          content:
            application/json:
              schema:
                type: object
  /import/cleanup:
    delete:
      tags:
        - Import/Export
      summary: Clean up old upload files
      description: Removes old temporary upload files
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                maxAge:
                  type: integer
                  default: 86400000
                  description: Maximum age in milliseconds
      responses:
        '200':
          description: Cleanup completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  max_age_hours:
                    type: number
                  result:
                    type: object
  /import/statistics:
    get:
      tags:
        - Import/Export
      summary: Get import statistics
      description: Retrieves comprehensive import statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Import statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  overview:
                    type: object
                    properties:
                      total_publications:
                        type: integer
                      total_authors:
                        type: integer
                      total_citations:
                        type: integer
                      recent_imports:
                        type: integer
                  imports_by_source:
                    type: array
                    items:
                      type: object
                  import_service:
                    type: object
                  etl_service:
                    type: object
  /import/export:
    get:
      tags:
        - Import/Export
      summary: Export publications
      description: Exports publications data in specified format
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json, csv]
        - name: filters
          in: query
          schema:
            type: object
      responses:
        '200':
          description: Export file download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    FileInfo:
      type: object
      properties:
        originalname:
          type: string
        filename:
          type: string
        size:
          type: integer
        mimetype:
          type: string
    ImportResult:
      type: object
      properties:
        total:
          type: integer
        processed:
          type: integer
        created:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              error:
                type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
...